+++ spec/support/data_store.rb
@@ -21,10 +21,10 @@ shared_examples_for 'a data store' do
     context 'with a name' do
       before do
         @data_store = Stoplight.data_store
-        Stoplight.data_store(data_store)
+        Stoplight.data_store = data_store
         Stoplight::Light.new(name) {}.run
       end
-      after { Stoplight.data_store(@data_store) }
+      after { Stoplight.data_store = @data_store }
 
       it 'includes the name' do
         expect(result).to include(name)